+++ packages/components/modal/src/modal-content.tsx
@@ -95,3 +95,6 @@ export const ModalContent = motionForwardRef<ModalContentProps, "section">(
     )
   },
 )
+
+ModalContent.displayName = "ModalContent"
+ModalContent.__ui__ = "ModalContent"