+++ packages/components/select/src/select.tsx
@@ -237,6 +237,9 @@ export const Select = forwardRef<SelectProps, "div">((props, ref) => {
   )
 })
 
+Select.displayName = "Select"
+Select.__ui__ = "Select"
+
 interface SelectFieldProps extends HTMLUIProps {}
 
 const SelectField = forwardRef<SelectFieldProps, "div">(
@@ -267,3 +270,6 @@ const SelectField = forwardRef<SelectFieldProps, "div">(
     )
   },
 )
+
+SelectField.displayName = "SelectField"
+SelectField.__ui__ = "SelectField"